+++ images.go
@@ -63,7 +63,7 @@ func listImages(images []*BaseImage) error {
 
 func listUserImages(images []UserImage) error {
 	for _, im := range images {
-		fmt.Printf("%s: (%d aliases)\n", im.Name, len(im.Aliases))
+		fmt.Printf("%s (%d envs)\n", im.Name, im.EnvCount)
 		fmt.Printf("  build time: %s\n", im.Labels["skeg.io/image/buildtime"])
 	}
 	return nil